pyproject.toml
[tool.poetry]
name = "llm-sql-agent-template"
version = "0.1.0"
description = ""
authors = ["Han Lee <lee.hanchung@gmail.com>"]
license = "LICENSE"
readme = "README.md"
packages = [{include = "sample_sql_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
tqdm = "^4.65.0"
duckdb = "^0.8.1"
duckdb-engine = "^0.7.3"
watchdog = "^3.0.0"
chainlit = "^0.6.1"
fire = "^0.5.0"
langchain = "^0.0.250"
debugpy = "1.6.7"
langchain-experimental = "^0.0.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.282"
pytest = "^7.4.0"
ipykernel = "^6.25.0"
ipywidgets = "^8.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
